Sign Language to Text Conversion for Dumb and Deaf
Sign language is one of the oldest and most natural form of language for
communication, but since most people do not know sign language and
interpreters are very difficult to come by we have come up with a real time
method using neural networks for fingerspelling based american sign
language. In our method, the hand is first passed through a filter and after
the filter is applied the hand is passed through a classifier which predicts the
class of the hand gestures. Our method provides 95.7 % accuracy for the 26
letters of the alphabet.